style(modeling) split cube out from cuboid code
